Lots of changes!
Added pieces, which are components meant to represent play pieces that go onto map tiles of the map.
Also heavily ramped up the SignalHub to support user signals. Being able to pass in arguments for user signals was quite a pain!
The SignalHub changes let the cursor dynamically update its appearance, depending on whether it's moved over a selectable object (such as a piece), or off of a selectable object.